requirements.txt
python-json-logger==2.0.7
pytz==2024.1
requests==2.31.0
requests-file==2.0.0
requests-toolbelt==1.0.0
python-decouple==3.8
urllib3==2.2.1
psycopg2-binary==2.9.9
SQLAlchemy==2.0.30
PureCloudPlatformClientV2==222.0.0